new commit with cookies and password encoding

+++ src/main/java/am/smartCode/jdbc/service/product/impl/ProductServiceImpl.java
@@ -1,6 +1,7 @@
 package am.smartCode.jdbc.service.product.impl;
 
-import am.smartCode.jdbc.exception.ValidationException;
+import am.smartCode.jdbc.exception.ProductNotFoundException;
+import am.smartCode.jdbc.exception.ProductValidationException;
 import am.smartCode.jdbc.model.Product;
 import am.smartCode.jdbc.repository.product.ProductRepository;
 import am.smartCode.jdbc.service.product.ProductService;
@@ -19,7 +20,7 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void createProduct(String category, String name, long price) throws SQLException {
-        productValidation(category, name,price);
+        productValidation(category, name, price);
         Product product = new Product();
         product.setCategory(category);
         product.setName(name);
@@ -30,14 +31,19 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public void updateProduct(Product product) throws SQLException {
+        if (product.getId() <= 0) {
+            throw new ProductValidationException(Message.INVALID_ID);
+        }
+        if (productRepository.get(product.getId()) == null) {
+            throw new ProductNotFoundException(Message.PRODUCT_NOT_FOUND);
+        }
+        productValidation(product.getCategory(), product.getName(), product.getPrice());
         Connection connection = productRepository.getConnection();
         connection.setAutoCommit(false);
         try {
-            productValidation(product.getCategory(), product.getName(),product.getPrice());
             productRepository.update(product);
             connection.commit();
-            System.out.println("Product is updated");
-        } catch (Throwable e) {
+        } catch (Exception e) {
             connection.rollback();
             connection.setAutoCommit(true);
         }
@@ -47,7 +53,10 @@ public class ProductServiceImpl implements ProductService {
     public Product getProduct(long id) throws SQLException {
         Connection connection = productRepository.getConnection();
         if (id <= 0) {
-            throw new ValidationException(Message.INVALID_ID);
+            throw new ProductValidationException(Message.INVALID_ID);
+        }
+        if (productRepository.get(id) == null) {
+            throw new ProductNotFoundException(Message.PRODUCT_NOT_FOUND);
         }
         return productRepository.get(id);
     }
@@ -56,14 +65,16 @@ public class ProductServiceImpl implements ProductService {
     public void deleteProduct(long id) throws SQLException {
         Connection connection = productRepository.getConnection();
         connection.setAutoCommit(false);
+        if (id <= 0) {
+            throw new ProductValidationException(Message.INVALID_ID);
+        }
+        if (productRepository.get(id) == null) {
+            throw new ProductNotFoundException(Message.PRODUCT_NOT_FOUND);
+        }
         try {
-            if (id <= 0) {
-                throw new ValidationException(Message.INVALID_ID);
-            }
             productRepository.delete(id);
             connection.commit();
-            System.out.println("Product is deleted");
-        } catch (Throwable e) {
+        } catch (Exception e) {
             connection.rollback();
             connection.setAutoCommit(false);
         }
@@ -72,7 +83,6 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public List<Product> getAllProducts() throws SQLException {
         Connection connection = productRepository.getConnection();
-        connection.setReadOnly(true);
         return productRepository.getAll();
     }
 
@@ -83,15 +93,15 @@ public class ProductServiceImpl implements ProductService {
         return productRepository.findProductsByName(name);
     }
 
-    private static void productValidation(String category, String name,long price) {
-        if (category.equals(null) || category.equals("")) {
-            throw new ValidationException(Message.BLANK_PRODUCT_CATEGORY);
+    private static void productValidation(String category, String name, long price) {
+        if (category == null || category.isEmpty()) {
+            throw new ProductValidationException(Message.BLANK_PRODUCT_CATEGORY);
         }
-        if (name.equals(null) || name.equals("")) {
-            throw new ValidationException(Message.BLANK_PRODUCT_NAME);
+        if (name == null || name.isEmpty()) {
+            throw new ProductValidationException(Message.BLANK_PRODUCT_NAME);
         }
         if (price <= 0) {
-            throw new ValidationException(Message.INVALID_PRICE);
+            throw new ProductValidationException(Message.INVALID_PRICE);
         }
     }
 }
